swiss_TLM_api/requirements.txt
rtree==1.0.1
fiona==1.8.22
grequests==0.6.0
shapely==1.8.5
Flask==2.3.2
gunicorn==22.0.0
flask-cors==4.0.1
gdown==4.5.4
pillow==10.4.0